refactor(bed-data-fetcher): separate shared types
